feat: alter user id

+++ models/models.py
@@ -88,7 +88,7 @@ class BotUser(Base):
     __tablename__ = "bot_users"
 
     id = Column(Integer, primary_key=True, index=True)
-    telegram_id = Column(Integer, unique=True, index=True)
+    telegram_id = Column(BigInteger, unique=True, index=True)
     username = Column(String, unique=True, index=True)
     first_name = Column(String)
     last_name = Column(String, nullable=True)
